completion parser is fairly complete and ready to be merged with master
