+++ src/main/java/org/aesh/command/impl/AeshCommandRuntime.java
@@ -33,9 +33,7 @@ import org.aesh.command.CommandResolver;
 import org.aesh.command.CommandResult;
 import org.aesh.command.Executor;
 import org.aesh.command.impl.internal.ProcessedCommand;
-import org.aesh.command.impl.parser.CommandLineCompletionParser;
 import org.aesh.command.impl.parser.CommandLineParser;
-import org.aesh.command.impl.parser.ParsedCompleteObject;
 import org.aesh.command.invocation.CommandInvocationBuilder;
 import org.aesh.command.parser.CommandLineParserException;
 import org.aesh.command.result.ResultHandler;
@@ -254,20 +252,15 @@ public class AeshCommandRuntime<C extends Command, CI extends CommandInvocation>
 
             try (CommandContainer commandContainer = commandResolver.resolveCommand(completeOperation.getBuffer())) {
 
-                CommandLineCompletionParser completionParser = commandContainer
-                        .getParser().getCompletionParser();
-
-                ParsedCompleteObject completeObject = completionParser
-                        .findCompleteObject(completeOperation.getBuffer(),
-                                completeOperation.getCursor());
-                completeObject.getCompletionParser().injectValuesAndComplete(completeObject,
-                        completeOperation, invocationProviders);
-            } catch (CommandLineParserException e) {
+                commandContainer.getParser().complete(completeOperation, invocationProviders);
+            }
+            catch (CommandLineParserException e) {
                 LOGGER.warning(e.getMessage());
-            } catch (CommandNotFoundException ignored) {
-            } catch (Exception ex) {
-                LOGGER.log(Level.SEVERE,
-                        "Runtime exception when completing: "
+            }
+            catch (CommandNotFoundException ignored) {
+            }
+            catch (Exception ex) {
+                LOGGER.log(Level.SEVERE, "Runtime exception when completing: "
                         + completeOperation, ex);
             }
         }
