+++ src/main/java/org/aesh/command/impl/internal/ProcessedOption.java
@@ -135,6 +135,13 @@ public final class ProcessedOption {
             return null;
     }
 
+    public String getLastValue() {
+        if(values.size() > 0)
+            return values.get(values.size()-1);
+        else
+            return null;
+    }
+
     public List<String> getValues() {
         return values;
     }
